requirements.txt
Django==4.0.3
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-taggit==2.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
environs==9.5.0
pillow